Changed the code over to ServerPlayerEntityMixin like creepycre said, i dunno if i did it right? but everything works fine so i think so????
