+++ src/main/java/org/dimdev/dimdoors/mixin/PlayerEntityMixin.java
@@ -25,10 +25,6 @@ import java.util.Random;
 
 @Mixin(value = PlayerEntity.class, priority = 900)
 public abstract class PlayerEntityMixin extends LivingEntity {
-	private static final int RANDOM_ACTION_BOUND = 400;
-	private static final int CHANCE_TO_DECREASE_ARMOR_DURABILITY = 20;
-	private static final int CHANCE_TO_REPLACE_ITEMSLOT_WITH_UNRAVLED_FABRIC = 10;
-	Random random = new Random();
 
 	@Shadow
 	public abstract void incrementStat(Identifier stat);
@@ -37,53 +33,7 @@ public abstract class PlayerEntityMixin extends LivingEntity {
 		super(entityType, world);
 	}
 
-	@Inject(method = "tick", at = @At("HEAD"), cancellable = true)
-	public void mobTickMixin(CallbackInfo ci) {
-		if (PlayerModifiersComponent.getFray(this) >= 125) {
-			if (random.nextInt(RANDOM_ACTION_BOUND) == 0) {
-				doRandomFunction(this);
-			}
-		}
-	}
 
-	private void doRandomFunction(LivingEntity player) {
-		switch (random.nextInt(2)) {
-			case 0:
-				decreaseArmorDurability((PlayerEntity) player);
-				break;
-			case 1:
-				addRandomUnravledFabric((PlayerEntity) player);
-				break;
-			default:
-		}
-
-	}
-
-	private void addRandomUnravledFabric(PlayerEntity player) {
-		if(PlayerModifiersComponent.getFray(player) < DimensionalDoorsInitializer.getConfig().getPlayerConfig().fray.unravledFabricInInventoryFray)
-			return;
-		if(random.nextInt(CHANCE_TO_REPLACE_ITEMSLOT_WITH_UNRAVLED_FABRIC) != 0)
-			return;
-
-		int slot = random.nextInt(player.getInventory().main.size());
-
-		if(!player.getInventory().main.get(slot).isEmpty() && ! (player.getInventory().main.get(slot).getItem() == ModItems.UNRAVELLED_FABRIC))
-			return;
-		if (player.getInventory().main.get(slot).getCount() >= 64)
-			return;
-		player.getInventory().main.set(slot, new ItemStack(ModItems.UNRAVELLED_FABRIC, 1 + player.getInventory().main.get(slot).getCount()));
-
-	}
-
-	private void decreaseArmorDurability(PlayerEntity player) {
-		if(PlayerModifiersComponent.getFray(player) < DimensionalDoorsInitializer.getConfig().getPlayerConfig().fray.unravledFabricInInventoryFray)
-			return;
-		for (int i = 0; i < player.getInventory().armor.size(); i++)
-			if (random.nextInt(CHANCE_TO_DECREASE_ARMOR_DURABILITY) == 0)
-				player.getArmorItems().forEach((itemStack) -> {
-					itemStack.setDamage(itemStack.getDamage() + 1);
-				});
-	}
 
 	@Inject(method = "handleFallDamage", at = @At("HEAD"), cancellable = true)
 	public void handleLimboFallDamage(float fallDistance, float damageMultiplier, DamageSource damageSource, CallbackInfoReturnable<Boolean> cir) {
